requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Deprecated==1.2.14
Flask==3.0.2
fonttools==4.49.0
frozenlist==1.4.1
future==0.18.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
line-bot-sdk==3.9.0
MarkupSafe==2.1.5
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.0
waitress==3.0.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
